change table names

+++ src/main/java/com/massonus/flavourflow/entity/PendingUser.java
@@ -13,5 +13,5 @@ public class PendingUser {
     @Id
     private Long id;
 
-    private Integer telegramId;
+    private Long telegramId;
 }
